Refactored testing-lib, added functionality and removed unused one.
