+++ utils/test-utils/src/main/java/com/microsoft/device/dualscreen/testing/DeviceRotation.kt
