+++ utils/test-utils/src/main/java/com/microsoft/device/dualscreen/testing/ForceClick.kt
@@ -4,7 +4,7 @@
  *
  */
 
-package com.microsoft.device.dualscreen.utils.test
+package com.microsoft.device.dualscreen.testing
 
 import android.view.View
 import androidx.test.espresso.UiController
@@ -14,8 +14,19 @@ import androidx.test.espresso.matcher.ViewMatchers.isEnabled
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.allOf
 
+
+/**
+ * Returns an action that clicks the view without checking the coordinates on the screen.
+ * Seems that ViewActions.click() finds coordinates of the view on the screen, and then performs the tap on the coordinates.
+ * Seems that changing the screen rotations affects these coordinates and ViewActions.click() throws exceptions.
+ */
+fun forceClick(): ViewAction = ForceClick()
+
 /**
  * Click action without checking the coordinates for the view on the screen. when device is rotated.
+ * ViewActions.click() finds coordinates of the view on the screen, and then performs
+ * the tap on the coordinates.
+ * It seems that changing the screen rotations affects these coordinates and ViewActions.click() throws exceptions.
  */
 class ForceClick : ViewAction {
     override fun getConstraints(): Matcher<View> {
